Add `findpid` alias for fuzzy incremental PID search Replace FZF with FZY tool for MacOS/Debian Add FZY keybindings for ZSH Update README.md Fix check all the prereqs on package install (!!!)

+++ packages/osx/zsh/zsh.sh
@@ -7,9 +7,15 @@ function install_package() {
 
   brew install zsh
 
-  brew unlink fzf
-  brew install fzf
-  /usr/local/opt/fzf/install
+  if [ -f /usr/local/opt/fzf/uninstall ]; then
+    ./usr/local/opt/fzf/uninstall
+    brew unlink fzf
+    brew uninstall --force fzf
+  fi
+
+  brew tap jhawthorn/fzy
+  brew install fzy
+  brew link fzy
 
   brew install thefuck
